bitcoinpricetracker/requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
boto==2.49.0
boto3==1.20.31
botocore==1.23.31
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cors==1.0.1
cryptography==35.0.0
debugpy==1.5.1
decorator==5.1.0
distlib==0.3.3
Django==3.2.8
django-braces==1.15.0
django-cleanup==5.2.0
django-cors-headers==3.6.0
django-cors-middleware==1.5.0
django-crispy-forms==1.13.0
django-crontab==0.7.1
django-dbbackup==1.80.1
django-extensions==3.1.5
django-filter==21.1
django-google-maps==0.12.4
django-modal-forms==2.0
django-oauth-toolkit==0.10.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-simple-sso==1.1.0
django-versatileimagefield==2.2
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
dnspython==2.1.0
drf-extra-fields==3.2.1
drf-flex-fields==0.9.6
drf-yasg==1.20.0
dropbox==11.25.0
entrypoints==0.3
environ==1.0
et-xmlfile==1.1.0
filelock==3.3.1
Flask==2.1.2
future==0.18.2
gevent==21.8.0
graphene==3.0
graphene-django==3.0.0b7
graphql-core==3.1.7
graphql-relay==3.1.5
greenlet==1.1.2
h11==0.13.0
idna==3.3
inflection==0.5.1
ipykernel==6.6.1
ipython==7.31.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jupyter-client==7.1.0
jupyter-core==4.9.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mongojoin==1.0.0
mysqlclient==2.1.0
Naked==0.1.31
nest-asyncio==1.5.4
numpy==1.22.0
oauthlib==1.0.3
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==21.3
pandas==1.3.5
parso==0.8.3
phonenumbers==8.12.18
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
ply==3.11
promise==2.3
prompt-toolkit==3.0.24
pycparser==2.20
Pygments==2.11.2
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==3.0.6
PyPDF2==1.26.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-magic==0.4.24
python-magic-bin==0.4.14
python-utils==2.5.6
pytz==2021.3
pywin32==303
PyYAML==6.0
pyzmq==22.3.0
razorpay==1.2.0
requests==2.26.0
requests-file==1.5.1
ruamel.yaml==0.17.17
s3transfer==0.5.0
shortuuid==1.0.8
simplejson==3.17.6
six==1.16.0
sqlparse==0.2.4
stone==3.2.1
text-unidecode==1.3
tldextract==3.1.2
tornado==6.1
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.7
utils==1.0.1
uvicorn==0.18.2
validators==0.18.2
virtualenv-clone==0.5.7
wcwidth==0.2.5
webservices==0.7
Werkzeug==2.1.2
whitenoise==5.3.0
win32-setctime==1.0.3
zope.event==4.5.0
zope.interface==5.4.0
